chore: restructure mods
