+++ crypto/src/lib.rs
@@ -1,7 +1,5 @@
 #![allow(clippy::too_many_arguments)]
 
-pub mod blind_rotation;
-pub mod distribution;
-pub mod lwe;
-pub mod rgsw;
-pub mod rlwe;
+pub mod core;
+pub mod scheme;
+pub mod util;
