+++ math/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod decomposer;
 pub mod distribution;
-pub mod misc;
 pub mod modulus;
 pub mod poly;
 pub mod ring;
+pub mod util;
