+++ math/src/ring/prime/noisy.rs
@@ -96,7 +96,6 @@ impl RingOps for NoisyPrimeRing {
 mod test {
     use crate::{
         distribution::Sampler,
-        misc::test::assert_precision,
         modulus::Prime,
         poly::ffnt::test::{poly_mul_prec_loss, round_trip_prec_loss},
         ring::{
@@ -104,6 +103,7 @@ mod test {
             test::{test_poly_mul, test_round_trip},
             RingOps,
         },
+        util::test::assert_precision,
     };
     use rand::{distributions::Uniform, thread_rng};
