+++ math/src/util/scratch.rs
